feat: prepare base lending contracts

+++ contracts/CTokenStorage.sol
@@ -35,12 +35,12 @@ abstract contract CTokenStorage is ICToken {
      * @notice Maximum borrow rate that can ever be applied (.0005% / block)
      */
 
-    uint internal constant borrowRateMaxMantissa = 0.0005e16;
+    uint256 internal constant borrowRateMaxMantissa = 0.0005e16;
 
     /**
      * @notice Maximum fraction of interest that can be set aside for reserves
      */
-    uint internal constant reserveFactorMaxMantissa = 1e18;
+    uint256 internal constant reserveFactorMaxMantissa = 1e18;
 
     /**
      * @notice Administrator for this contract
@@ -65,47 +65,47 @@ abstract contract CTokenStorage is ICToken {
     /**
      * @notice Initial exchange rate used when minting the first CTokens (used when totalSupply = 0)
      */
-    uint internal initialExchangeRateMantissa;
+    uint256 internal initialExchangeRateMantissa;
 
     /**
      * @notice Fraction of interest currently set aside for reserves
      */
-    uint public reserveFactorMantissa;
+    uint256 public reserveFactorMantissa;
 
     /**
      * @notice Block number that interest was last accrued at
      */
-    uint public accrualBlockNumber;
+    uint256 public accrualBlockNumber;
 
     /**
      * @notice Accumulator of the total earned interest rate since the opening of the market
      */
-    uint public borrowIndex;
+    uint256 public borrowIndex;
 
     /**
      * @notice Total amount of outstanding borrows of the underlying in this market
      */
-    uint public totalBorrows;
+    uint256 public totalBorrows;
 
     /**
      * @notice Total amount of reserves of the underlying held in this market
      */
-    uint public totalReserves;
+    uint256 public totalReserves;
 
     /**
      * @notice Total number of tokens in circulation
      */
-    uint public totalSupply;
+    uint256 public totalSupply;
 
     /**
      * @notice Official record of token balances for each account
      */
-    mapping (address => uint) internal accountTokens;
+    mapping(address => uint256) internal accountTokens;
 
     /**
      * @notice Approved token transfer amounts on behalf of others
      */
-    mapping (address => mapping (address => uint)) internal transferAllowances;
+    mapping(address => mapping(address => uint256)) internal transferAllowances;
 
     /**
      * @notice Container for borrow balance information
@@ -113,8 +113,8 @@ abstract contract CTokenStorage is ICToken {
      * @member interestIndex Global borrowIndex as of the most recent balance-changing action
      */
     struct BorrowSnapshot {
-        uint principal;
-        uint interestIndex;
+        uint256 principal;
+        uint256 interestIndex;
     }
 
     /**
@@ -125,6 +125,5 @@ abstract contract CTokenStorage is ICToken {
     /**
      * @notice Share of seized collateral that is added to reserves
      */
-    uint public constant protocolSeizeShareMantissa = 2.8e16; //2.8%
-
+    uint256 public constant protocolSeizeShareMantissa = 2.8e16; //2.8%
 }
